backend/requirements.txt
# Framework y ORM
Django==4.2.7
django-tenant-schemas==1.11.0
djangorestframework==3.14.0
django-cors-headers==4.3.0

# Base de Datos
mysqlclient==2.2.0
redis==5.0.1

# Autenticación y Seguridad
djangorestframework-simplejwt==5.3.0
django-allauth==0.57.0
cryptography==41.0.7

# Tareas Asíncronas
celery==5.3.4
django-celery-beat==2.5.0

# Reportes y Utilidades
reportlab==4.0.6
openpyxl==3.1.2
Pillow==10.1.0
python-decouple==3.8

# Development
python-dotenv==1.0.0
